Make `SetValue` and `GetValue` templated.
`SetValue` and `GetValue` functions are now using template to specify
the type of the value to store. So now integers will be stored as
integeres and strings as strings and not not everything as strings.
Nevertheless Flow Graph node `SetValue` will convert the value to
string and `GetValue` will read the value as string.