Replace `bOverwrite` with `bEnableBackup` in saving.
I replaced `bOverwrite` parameter with `bEnableBackup` in `SaveDocument`
function. I though that you would most likely want to overwrite the file
and since we have other ways to tell whether the file exists or not
I replaced this option with `bEnableBackup` which will also save a
backup file which `LoadDocument` can use if the original file is missing.